mysite/requirements.txt
Django==1.9.8
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-mptt==0.8.7
django-user-accounts==1.3.1
djangorestframework==3.5.4
jsonfield==1.0.3
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.10.1
pinax-webanalytics==2.0.4
pytz==2016.10
